refactor(*): hook the events backup for selecting a character
