+++ EorzeanInfo/app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".views.FirstLaunchActivity" />
         <activity
-            android:name=".views.CharacterSelectActivity"
+            android:name=".views.CharacterChangeActivity"
             android:label="@string/title_activity_character_select"
             android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
