+++ EorzeanInfo/app/src/main/java/chesire/eorzeaninfo/views/CharacterChangeActivity.java
@@ -2,23 +2,22 @@ package chesire.eorzeaninfo.views;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-
-import java.util.ArrayList;
+import android.support.v7.widget.Toolbar;
+import android.view.View;
 
 import javax.inject.Inject;
 
+import butterknife.ButterKnife;
+import butterknife.OnClick;
 import chesire.eorzeaninfo.EorzeanInfoApp;
 import chesire.eorzeaninfo.R;
 import chesire.eorzeaninfo.classes.CharacterModel;
 import chesire.eorzeaninfo.interfaces.CharacterStorage;
 
-/**
- * Activity used for searching for a new character to use
- */
-public class CharacterSearchActivity extends AppCompatActivity
-        implements CharacterSearchFragment.CharacterSearchListener, CharacterSelectFragment.CharacterSelectListener {
-    private static final String CHARACTER_SEARCH_TAG = "CHARACTER_SEARCH_TAG";
+public class CharacterChangeActivity extends AppCompatActivity implements CharacterSelectFragment.CharacterSelectListener {
+
     private static final String CHARACTER_SELECT_TAG = "CHARACTER_SELECT_TAG";
 
     @Inject
@@ -27,25 +26,26 @@ public class CharacterSearchActivity extends AppCompatActivity
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_character_search);
+
+        setContentView(R.layout.activity_character_change);
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
 
         ((EorzeanInfoApp) getApplication()).getCharacterStorageComponent().inject(this);
+        ButterKnife.bind(this);
 
         if (savedInstanceState == null) {
             getSupportFragmentManager()
                     .beginTransaction()
-                    .add(R.id.character_search_container, CharacterSearchFragment.newInstance(), CHARACTER_SEARCH_TAG)
+                    .add(R.id.character_change_container, CharacterSelectFragment.newInstance(mCharacterStorage.getAllCharacters()), CHARACTER_SELECT_TAG)
                     .commit();
         }
     }
 
-    @Override
-    public void onCharactersFound(ArrayList<CharacterModel> models) {
-        getSupportFragmentManager()
-                .beginTransaction()
-                .replace(R.id.character_search_container, CharacterSelectFragment.newInstance(models), CHARACTER_SELECT_TAG)
-                .addToBackStack(null)
-                .commit();
+    @OnClick(R.id.character_select_fab)
+    void onFabClicked(View v) {
+        // go to the search fragment
+        Snackbar.make(v, "Replace with your own action", Snackbar.LENGTH_LONG).setAction("Action", null).show();
     }
 
     @Override
