--- EorzeanInfo/app/src/main/java/chesire/eorzeaninfo/views/CharacterSelectActivity.java
@@ -1,32 +1,0 @@
-package chesire.eorzeaninfo.views;
-
-import android.os.Bundle;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-
-import butterknife.ButterKnife;
-import butterknife.OnClick;
-import chesire.eorzeaninfo.R;
-
-public class CharacterSelectActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        setContentView(R.layout.activity_character_select);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        ButterKnife.bind(this);
-    }
-
-    @OnClick(R.id.character_select_fab)
-    void onFabClicked(View v) {
-        // go to the search fragment
-        Snackbar.make(v, "Replace with your own action", Snackbar.LENGTH_LONG).setAction("Action", null).show();
-    }
-
-}
