+++ EorzeanInfo/app/src/main/res/menu/menu_character_select.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="chesire.eorzeaninfo.views.CharacterSelectActivity">
+    tools:context="chesire.eorzeaninfo.views.CharacterChangeActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
